Calculate system dynamics sparse

This is a helper function that calculates the dynamics at time t of all the nodes in the system. It is seldom called on it's own but is called by other functions. This function comes in a version for sparse and dense matrices.

Calc_System_Dynamics_sparse(
  NodeStatus,
  ten_mat,
  damp_mat,
  kvect,
  dvect,
  mat_size,
  tstep = 1,
  non_empty_matrix,
  frctmultiplier = 1
)

Arguments

NodeStatus
A numeric matrix frame The current dynamics and forces experienced by the nodes. The node ID is not included This means the order of the matrix has to be fixed.

ten_mat
A numeric matrix. This matrix must be a zero matrix or an adjacency matrix of dimensions nxn where n is the number of nodes

damp_mat
Identical to the ten_mat. This is supposed to help reduce memory, But I don't think it does

kvect
A numeric vector of the spring stiffnesses

dvect
A numeric vector of the initial distances between the nodes

mat_size
a numeric scaler. The dimensions of the square matrix

tstep
A numeric value. The time step, measured in seconds, that will be used to calculate the new dynamic state

non_empty_matrix
A numeric matrix. This contains the node indexes in the adjacency matrix it has 4 columns the row, the column, the absolute index the transpose of the absolute index.

frctmultiplier
A numeric value. Used to set a multiplier on the friction value. Generally leave this alone.

Details

The output of the function is a matrix that is the next iteration of the NodeStatus dataframe the function recieves.