Create stabilised blocks

An internal function. This function is called by SETSe_bicomp and performs auto_SETSe on all the bi-connected components of the network. This function is rarely called directly.

Create_stabilised_blocks(
  g,
  OriginBlock,
  OriginBlock_number,
  force = "force",
  edge_name = "edge_name",
  k = "k",
  tstep = 0.1,
  tol = 1e-10,
  distance,
  max_iter,
  mass,
  sparse,
  sample = 100,
  static_limit = NULL,
  hyper_iters = 100,
  hyper_tol = 0.01,
  hyper_max = 30000,
  drag_min = drag_min,
  drag_max = drag_max,
  tstep_change = tstep_change,
  verbose = FALSE,
  bigraph = bigraph,
  balanced_blocks = balanced_blocks,
  noisey_termination = TRUE
)

Arguments

g
An igraph object

OriginBlock
An Igraph object. This is created by Create_balanced_blocks function

OriginBlock_number
An integer. this is the origin block chosen from the create_stable_blocks function. Usually this will be the largest block.

force
A character string. This is the node attribute that contains the force the nodes exert on the network.

edge_name
A character string. This is the edge attribute that contains the edge_name of the edges.

k
A character string. name of the spring constant variable

tstep
A numeric. The time interval used to iterate through the network dynamics.

tol
A numeric. The tolerance factor for early stopping.

distance
A character string. The edge attribute that contains the original/horizontal distance between nodes.

max_iter
An integer. The maximum number of iterations before stopping. Larger networks usually need more iterations.

mass
A numeric. This is the mass constant of the nodes in normalised networks this is set to 1.

sparse
Logical. Whether or not the function should be run using sparse matrices. must match the actual matrix, this could prob be automated

sample
Integer. The dynamics will be stored only if the iteration number is a multiple of the sample. This can greatly reduce the size of the results file for large numbers of iterations. Must be a multiple of the max_iter

static_limit
Numeric. The maximum value the static force can reach before the algorithm terminates early. This prevents calculation in a diverging system. The value should be set to some multiple greater than one of the force in the system. If left blank the static limit is twice the system absolute mean force.

hyper_iters
integer. The hyper parameter that determines the number of iterations allowed to find an acceptable convergence value.

hyper_tol
numeric. The convergence tolerance when trying to find the minimum value

hyper_max
integer. The maximum number of iterations that the setse will go through whilst searching for the minimum.

drag_min
integer. A power of ten. The lowest drag value to be used in the search

drag_max
integer. A power of ten. if the drag exceeds this value the tstep is reduced

tstep_change
numeric. A value between 0 and 1 that determines how much the time step will be reduced by default value is 0.5

verbose
Logical. This value sets whether messages generated during the process are supressed or not.

bigraph
A list. the list of biconnected components produced by the biconnected_components function. This function take a non trivial amount of time on large graphs so this pass through minimises the function being called.

balanced_blocks
A list

noisey_termination
Stop the process if the static force does not monotonically decrease.

Value

A dataframe with the height embeddings of the network

Details

This function isn't really supposed to be used apart from as a sub-routine of the SETSe biconnected component method.

See also