Calculate SETSe expanded version

This is a special case function which keeps the history of the network dynamics. It is useful for demonstrations. or Parametrizing difficult networks

SETSe_expanded(
  g,
  force = "net_generation",
  distance = "distance",
  edge_name = "edge_name",
  k = "k",
  tstep = 0.02,
  mass = 1,
  max_iter = 20000,
  coef_drag = 1,
  tol = 1e-06,
  sparse = FALSE,
  verbose = TRUE,
  two_node_solution = TRUE
)

Arguments

g
An igraph object. The network

force
A character string

distance
A character string. The name of the graph attribute that contains the graph distance

edge_name
A character string. This is the edge attribute that contains the edge_name of the edges.

k
A character string. This is k for the moment don't change it.

tstep
A numeric. The time in seconds that elapses between each iteration

mass
A numeric. The mass in kg of the nodes, this is arbitrary and commonly 1 is used.

max_iter
An interger. The maximum nuber of iterations before terminating the simulation

coef_drag
A numeric. A multplier used to tune the damping. Generally no need to twiddle

tol
A numeric. Early termination. If the dynamics of the nodes fall below this value the algortihm will be classed as "converged" and the simulation terminates.

sparse
Logical. Whether or not the function should be run using sparse matrices. must match the actual matrix, this could prob be automated

verbose
Logical value. Whether the function should output messages or run quietly.

two_node_solution
Logical. The newton-raphson algo is used to find the correct angle