Auto SETSe

Uses a grid search and a binary search to find appropriate convergence conditions.

auto_SETSe2(
  g,
  force = "force",
  distance = "distance",
  edge_name = "edge_name",
  k = "k",
  tstep = 0.02,
  mass = 1,
  max_iter = 1e+05,
  tol = 0.002,
  sparse = FALSE,
  hyper_iters = 100,
  hyper_tol = 0.01,
  hyper_max = 30000,
  lower_drag = 0.01,
  upper_drag = 10,
  tstep_change = 0.1,
  step_size = NA,
  sample = 100,
  static_limit = NULL,
  verbose = FALSE,
  include_edges = TRUE
)

Arguments

g
An igraph object

force
A character string. This is the node attribute that contains the force the nodes exert on the network.

distance
A character string. The edge attribute that contains the original/horizontal distance between nodes.

edge_name
A character string. This is the edge attribute that contains the edge_name of the edges.

k
A character string. This is k for the moment don't change it.

tstep
A numeric. The time interval used to iterate through the network dynamics.

mass
A numeric. This is the mass constant of the nodes in normalised networks this is set to 1.

max_iter
An integer. The maximum number of iterations before stopping. Larger networks usually need more iterations.

tol
A numeric. The tolerance factor for early stopping.

sparse
Logical. Whether or not the function should be run using sparse matrices. must match the actual matrix, this could prob be automated

hyper_iters
integer. The hyper parameter that determines the number of iterations allowed to find an acceptable convergence value.

hyper_tol
numeric. The convergence tolerance when trying to find the minimum value

hyper_max
integer. The maximum number of iterations that the setse will go through whilst searching for the minimum.

step_size
numeric. The hyper parameter that determines the log ratio search step size for auto convergence

sample
Integer. The dynamics will be stored only if the iteration number is a multiple of the sample. This can greatly reduce the size of the results file for large numbers of iterations. Must be a multiple of the max_iter

static_limit
Numeric. The maximum value the static force can reach before the algorithm terminates early. This prevents calculation in a diverging system. The value should be set to some multiple greater than one of the force in the system. If left blank the static limit is the system absolute mean force.

inlcude_edges
logical. An optional variable on wehther to calculate the edge tension and strain. Default is TRUE. included for ease of integration into the bicomponent functions.

Value

A list of four elements. A data frame with the height embeddings of the network, a data frame of the edge embeddings, the convergence dynamics dataframe for the network as well as the search history for convergence criteria of the network

Details

This function is pretty useful, it takes advantage of the linear relationship between the timestep and the coefficient of drag to search along the log linear line formed by tstep/coef_drag to find the convergence conditions.

See also