Prepare categorical features for embedding

This function prepares a binary network for SETSe projection.

prepare_SETSe_binary(
  g,
  node_names,
  k = NULL,
  force_var,
  sum_to_one = TRUE,
  distance = 1
)

Arguments

g
an igraph object

node_names
a character string. A vertex attribute which contains the node names.

k
The spring constant. This value is either a numeric value giving the spring constant for all edges or NULL. If NULL is used the k value will not be added to the network. This is useful k is made through some other process.

force_var
A node attribute. This is used as the force variable, it must be a character or factor

sum_to_one
Logical. whether the total positive force sums to 1, if FALSE the total is the sum of the positive cases

distance
a positive numeric value. The default is 1

Value

A network with the correct edge and node attributes for the embeddings process.

Details

The network takes in an igraph object and produces an undirected igraph object that can be used with the embedding functions.

The function converts the attribute force_var to a numeric force value. It also creates the edge attribute 'k' unless k=NULL. The purpose of the function is to easily be able to project categorical features using SETSe. The function creates new variables where each variable represents one level of the categorical variables. For embedding only n-1 of the levels are needed.

The function creates several variables

force_: Vertex attributes representing the force produced by each node for each categorical value, there will be n of these variables representing each level of the categorical values. The variable names will be the the name of the variable and the name of the level seperated by and underscore. For example, with a variable group and levels A and B, the created force variables will be "group_A" and "group_B" The sum of these variables will be 0

k: The spring constant representing the stiffness of the spring.

edge_name: the name of the edges. it takes the form "from_to" where "from" is the origin node and "to" is the destination node using the as_data_frame function from igraph

See also

Examples

set.seed(234) #set the random see for generating the network
g <- generate_peels_network(type = "E")
embeddings <- g %>%
#prepare the network for a binary embedding
prepare_SETSe_binary(., node_names = "name", k = 1000, 
                     force_var = "class") %>%
#embed the network using auto setse
  SETSe_auto(., force = "group_A")

#> Error in mutate(., dplyr::across(.fns = ~{    .x - mean(.x)})): could not find function "mutate"